+++ docs/release-notes/change-log.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+* Add a substreams `live back filler` once substreams tier1 is requested in `production mode`. 
+  It enables, create `substreams cache` when block are processed live on tier1. 
+
 ### Remote Code Generation
 
 The `substreams init` command now fetches a list of available 'code generators' to "https://codegen.substreams.dev".